Add files via upload
Initial commit. A Generic Switch class to add and control Devices classes Switch is a generic class which can add any number of devices, get their state, control their state and get and update various parameters of the Devices class. Switch is like a mobile app model for smart connected home.

Switch class has the logic and control code for the switch, Devices class is the base class for all the devices and Television and Refrigerator are the derived classes of Devices class